requirements/common.txt
# ^^^^^^^^^^^^^^^^^^^^^^^^
# DJANGO & EXTENSIONS
# Also requires: None
# ^^^^^^^^^^^^^^^^^^^^^^^^
Django==2.0.2
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
django-extensions==2.0.7
django-model-utils==3.1.1
drf-extensions==0.3.1

# ^^^^^^^^^^^^^^^^^^^^^^
# DATABASE MODULES
# ^^^^^^^^^^^^^^^^^^^^^^
psycopg2==2.7.4

# ^^^^^^^^^^^^^^^^^^^^^^
# TESTING
# ^^^^^^^^^^^^^^^^^^^^^^
factory-boy==2.10.0
Faker==0.8.13

# ^^^^^^^^^^^^^^^^^^^^^^
# STANDALONE
# ^^^^^^^^^^^^^^^^^^^^^^
django-cors-headers==2.4.0
django-shortuuidfield==0.1.3
requests==2.18.4
gunicorn==19.7.1
coreapi==2.3.3

# ^^^^^^^^^^^^^^^^^^^^^^
# INTEGRATIONS
# ^^^^^^^^^^^^^^^^^^^^^^
